refactor: extract useAuthFileHandlers hook (#3501 Phase 1j)
